Move files to new naming std
