+++ os/hal/ports/NRF51/NRF51822/hal_adc_lld.h
