+++ os/hal/ports/NRF51/NRF51822/hal_ext_lld.h
