+++ os/hal/ports/NRF51/NRF51822/hal_gpt_lld.h
