+++ os/hal/ports/NRF51/NRF51822/hal_i2c_lld.c
