+++ os/hal/ports/NRF51/NRF51822/hal_pal_lld.c
