+++ os/hal/ports/NRF51/NRF51822/hal_rng_lld.c
