+++ os/hal/ports/NRF51/NRF51822/hal_serial_lld.c
