+++ os/hal/ports/NRF51/NRF51822/hal_spi_lld.c
