+++ os/hal/ports/NRF51/NRF51822/hal_st_lld.h
