+++ os/hal/ports/NRF51/NRF51822/hal_wdg_lld.h
